requirements.txt
python-dotenv
requests
playsound
SpeechRecognition
openai
pyaudio